Revamped bug and suggestion systems
Moved restart command from bot category to developer category.
The bug and suggestion system has been thoroughly revamped. You can report a bug or submit a suggestion either on the bot or on the website.
Either way you do it, you will still be notified whenever a developer views the bug or suggestion on the bot or on the website.